feat(01-02): retype money/quantity columns and make finance.py Decimal-native
Task 1: money columns -> Numeric(12,2), saleable-quantity columns -> Integer,
across sku/order_item/order_item_tax/account/credit_note_item/inventory/
sku_batch/order_item_batch/inventory_transaction models. Request schemas
OrderItemCreate.quantity and CreditNoteItemCreate.quantity narrowed to int
so a fractional quantity is rejected with 422 at the API boundary.
Geospatial/dimension columns (weight, length/width/height_cm, lat/lng) left
Float per plan.

Plan boundary correction (found during Task 1 verification, directed by
coordinator): the plan assumed finance.py's service-layer arithmetic would
tolerate the type change untouched, deferring all finance.py work to 01-03.
That assumption was wrong -- _round_money() converted its return value to
float, which raised TypeError when subtracted from the now-Decimal
intermediate values upstream. Fixed by making _round_money Decimal-in/
Decimal-out (Decimal(str(x)), never Decimal(float) or float() on a Decimal
operand, to avoid reintroducing binary-float artifacts). calculate_order_totals,
calculate_order_item_totals, _order_item_taxable_value, _order_item_tax_amount
and calculate_credit_note_totals now flow Decimal through unchanged.
calculate_order_outstanding additionally routes each order.payments amount
through _round_money before accumulating, since a payment appended to
order.payments within the same transaction (e.g. create_payment, pre-flush)
still holds the raw Python value assigned at construction rather than the
Decimal SQLAlchemy returns after a round-trip through the Numeric column.
The GST formula itself (inclusive_value * (total_rate / 100)) is unchanged --
out of scope per coordinator direction, a CA-facing decision for later.

test_finance.py tightened from pytest.approx(float) to exact Decimal
comparison now that the arithmetic is exact end-to-end (CONTEXT.md
explicitly allows this: "tightening pytest.approx() where exactness now
holds").

accounting.py still constructs Account(amount=round(float(...), 2)) --
a float assignment into a Numeric column -- left untouched per coordinator
direction; deferred to 01-03 along with order.py and _normalize_state/
_is_inter_state.

Suite: 20 passed / 0 failed (baseline maintained).

+++ app/models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Enum
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, ForeignKey, Enum
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 
@@ -10,7 +10,7 @@ class Account(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     transaction_reference = Column(String, unique=True, nullable=False)
     payment_mode = Column(Enum(PaymentMode, name="payment_mode"), nullable=True)
     collected_by_id = Column(Integer, ForeignKey("employees.id"), nullable=True)
