+++ app/models/credit_note_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.db.base import Base
@@ -9,7 +9,7 @@ class CreditNoteItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     credit_note_id = Column(Integer, ForeignKey("credit_notes.id"), nullable=False)
     sku_id = Column(Integer, ForeignKey("skus.id"), nullable=False)
-    quantity = Column(Float, nullable=False)
-    unit_price = Column(Float, nullable=False)
+    quantity = Column(Integer, nullable=False)
+    unit_price = Column(Numeric(12, 2), nullable=False)
 
     credit_note = relationship("CreditNote", back_populates="items")
